Use RTL's waitFor() utility to get rid of "act()" related warnings
Reference:
https://kentcdodds.com/blog/write-fewer-longer-tests

Wrap assertions in waitFor() so that the "act()" warnings don't happen.
All of our requests are mocked so return immediately, so I think this
is fine to do.
Except in the first test, where you don't really have anything to do
with the result of the async call, so just waiting and ignoring the
result.